+++ TerranEngine/src/Project/ProjectSerializer.cpp
@@ -78,8 +78,8 @@ namespace TerranEngine
         {
             auto physicsSettings = node["Physics"];
             m_Project->m_PhysicsSettings.Gravity = physicsSettings["Gravity"].as<glm::vec2>();
-            m_Project->m_PhysicsSettings.VelocityIterations = physicsSettings["VelocityIterations"].as<float>();
-            m_Project->m_PhysicsSettings.PositionIterations = physicsSettings["PositionIterations"].as<float>();
+            m_Project->m_PhysicsSettings.VelocityIterations = physicsSettings["VelocityIterations"].as<uint32_t>();
+            m_Project->m_PhysicsSettings.PositionIterations = physicsSettings["PositionIterations"].as<uint32_t>();
             m_Project->m_PhysicsSettings.PhysicsTimestep = physicsSettings["PhysicsTimestep"].as<float>();
 
             int i = 0;
